price-optimser-backend/requirements.txt
Django>=3.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
